backend/requirements.txt
# Web API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# LLM & local embeddings
sentence-transformers>=2.2.0
groq>=0.4.0
httpx>=0.26.0

# Vector search
faiss-cpu>=1.7.4
numpy>=1.24.0

# Token counting for chunking (aligned with OpenAI tokenizers)
tiktoken>=0.5.0

# Scraping (Phase 1)
beautifulsoup4>=4.12.0
playwright>=1.40.0
